refactor: extract NE mgmt actions to nuvlaedge_k8s module
